Add label mgmt pages to frontend, update frontend services

+++ Frontend/src/app/services/label.service.ts
@@ -6,6 +6,8 @@ import { HttpService } from './http.service';
 export interface Label {
   id: string;
   title: string;
+  createdAt: string;
+  updatedAt: string;
 }
 
 @Injectable({
@@ -33,6 +35,20 @@ export class LabelService {
     }).then(response => response.data);
   }
 
+  update(labelId: string, props: Partial<Label>) {
+    const url = this.utilService.getBase() + 'labels/' + labelId + this.utilService.getTokenQuery();
+
+    return this.httpService.request({
+      method: 'put',
+      url,
+      data: props
+    }).then(response => {
+      this.events.publish('label:updated');
+
+      return response.data;
+    });
+  }
+
   createBulk(data) {
     const url = this.utilService.getBase() + 'labels/' + this.utilService.getTokenQuery();
 
@@ -47,12 +63,29 @@ export class LabelService {
     });
   }
 
-  remove(data) {
+  // Removes label from a single associated recipe
+  removeFromRecipe(labelId: string, recipeId: string) {
+    if (!labelId || !recipeId) throw new Error(`Invalid recipeId or labelId`);
+
     const url = this.utilService.getBase() + 'labels/' + this.utilService.getTokenQuery()
-                + '&labelId=' + data.id + '&recipeId=' + data.recipeId;
+                + '&labelId=' + labelId + '&recipeId=' + recipeId;
 
     return this.httpService.request({
       method: 'delete',
+      url
+    }).then(response => {
+      this.events.publish('label:deleted');
+
+      return response.data;
+    });
+  }
+
+  // Deletes label and removes from all associated recipes
+  delete(labelIds: string[]) {
+    const url = this.utilService.getBase() + 'labels/delete' + this.utilService.getTokenQuery() + '&labelIds=' + labelIds.join(',');
+
+    return this.httpService.request({
+      method: 'post',
       url
     }).then(response => {
       this.events.publish('label:deleted');
